+++ src/ui/modal_renderers_plan.rs
@@ -16,6 +16,7 @@ use super::i18n::Msg;
 use super::modal_renderers::{render_plan_modal_wrapper_styled, ModalIcon};
 use super::modal_renderers_input::render_input_plan_modal;
 use super::modal_renderers_stash::render_stash_planning;
+use super::modals::worktree::*;
 use super::modals::*;
 use super::theme::{self as theme_mod, theme};
 use super::{KagiApp, MONO_FONT};